Fix neovim-installed condition

+++ install.sh
@@ -11,6 +11,14 @@ SRCS="
   bin
 "
 
+SRCS="
+  .config
+  .gitignore
+  .tmux.conf
+  .profile
+  bin
+"
+
 cp -rv $SRCS ~/
 
 # git-config  special treatment: preserve user info
@@ -33,7 +41,7 @@ fi
 # neovim
 #
 nvim_dir=$HOME/nvim-linux-x86_64
-if ! type nvim >/dev/null 2>&1 | grep -Eq "/local/|$HOME" && ! [[ -d $nvim_dir ]] ; then
+if ! type nvim >/dev/null 2>&1 && ! type nvim >/dev/null 2>&1 | grep -Eq "/local/|$HOME" && ! [[ -d $nvim_dir ]] ; then
     cd /tmp
     rm nvim-linux-x86_64.tar.gz
     curl -L -o /tmp/nvim-linux-x86_64.tar.gz https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
